[util] Move runtime util to its own package
- to avoid the import cycle, since every package imports util, we
copied GetCallerPackage in log package
- [test] add docker-compose based test for 1.8 and 1.9
