+++ log/logger.go
@@ -4,10 +4,9 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"runtime"
-	"strings"
 	"sync"
 
+	"github.com/dyweb/gommon/util/runtimeutil"
 	"github.com/pkg/errors"
 )
 
@@ -107,7 +106,7 @@ func (log *Logger) NewEntryWithPkg(pkgName string) *Entry {
 // TODO: this is better than do filter in logger since we can apply the logging to each entry
 func (log *Logger) RegisterPkg() *Entry {
 	fields := make(map[string]string, 1)
-	pkg := getCallerPackage(2)
+	pkg := runtimeutil.GetCallerPackage(2)
 	fields["pkg"] = pkg
 	e := &Entry{
 		Logger:     log,
@@ -130,16 +129,4 @@ func (log *Logger) PrintEntries() {
 	for pkg := range log.Entries {
 		fmt.Println(pkg)
 	}
-}
-
-// FIXME: it should be in util package, but we put it here to avoid import cycle
-func getCallerPackage(skip int) string {
-	pc, _, _, ok := runtime.Caller(skip)
-	if !ok {
-		return "unknown"
-	}
-	fn := runtime.FuncForPC(pc)
-	fnName := fn.Name()
-	lastDot := strings.LastIndex(fnName, ".")
-	return fnName[:lastDot]
 }
